enhancement(json_parser transform): Enhance instrumentation
